+++ packages/expo-ui/src/swift-ui/Image/index.tsx
@@ -1,4 +1,5 @@
 import { requireNativeView } from 'expo';
+import { ColorValue } from 'react-native';
 import { type SFSymbol } from 'sf-symbols-typescript';
 
 import { type ViewEvent } from '../../types';
@@ -19,7 +20,7 @@ export interface ImageProps extends CommonViewModifierProps {
    * The color of the system image.
    * Can be a color name like '#ff00ff', 'red', 'blue', etc.
    */
-  color?: string;
+  color?: ColorValue;
   /**
    * The variable value for SF Symbols with variable color support.
    * Can be a number between 0.0 and 1.0.